chore: moved slang table to learn rootnav
