+++ react_main/src/pages/Learn/Learn.jsx
@@ -19,8 +19,67 @@ import { GameTypes } from "../../Constants";
 
 import "../../css/play.css";
 
+import {
+  Paper,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+} from "@mui/material";
+import { slangList } from "../../constants/slangList";
+
 export default function Learn(props) {
   const defaultGameType = "Mafia";
+
+  const slangTableRows = Object.keys(slangList).map((key) => {
+    let { definition, emoji } = slangList[key];
+    if (Array.isArray(emoji)) {
+      emoji = emoji.join(", ");
+    }
+
+    return {
+      term: key,
+      definition,
+      emoji,
+    };
+  });
+
+    // TODO [MUI]: modify the theme rather than using 'sx', for consistency (across all components)
+    const slangTable = (
+      <TableContainer component={Paper}>
+        <Table aria-label="a dense table">
+          <TableHead>
+            <TableRow>
+              <TableCell align="center" sx={{ fontWeight: "bold" }}>
+                Term
+              </TableCell>
+              <TableCell align="center" sx={{ fontWeight: "bold" }}>
+                Explanation
+              </TableCell>
+              <TableCell align="center" sx={{ fontWeight: "bold" }}>
+                'Additions'
+              </TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {slangTableRows.map((row) => (
+              <TableRow
+                key={row.name}
+                sx={{ "&:last-child td, &:last-child th": { border: 0 } }}
+              >
+                <TableCell component="th" scope="row" align="center">
+                  {row.term}
+                </TableCell>
+                <TableCell align="center">{row.definition}</TableCell>
+                <TableCell align="center">{row.emoji}</TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+    );
 
   const location = useLocation();
   const params = new URLSearchParams(location.search);
@@ -85,6 +144,23 @@ export default function Learn(props) {
 
           <Route render={() => <Redirect to="/play" />} />
         </Switch>
+        <div className="heading">Terminology (mafia slang)</div>
+        <div className="paragraph">
+          Below lies the full list of terms automatically detected by the game.
+          <br />
+          <br />
+          If you would like to improve one of the explanations (or even the
+          emoji pool) or add a new term,{" "}
+          <strong>
+            please consider contributing through our Feedback form / Discord /
+            Github Repo
+          </strong>{" "}
+          / etc.
+          <br />
+          <br />
+          It's up to us to keep it fresh and relevant.
+        </div>
+        <div className="paragraph">{slangTable}</div>
       </div>
     </>
   );
